+++ examples/aes.rs
@@ -10,7 +10,7 @@ use cortex_m_rt::entry;
 #[allow(unused_imports)]
 use hal::prelude::*;
 #[allow(unused_imports)]
-use lpc55s6x_hal as hal;
+use lpc55_hal as hal;
 
 use aes_soft::block_cipher_trait::generic_array::GenericArray;
 use aes_soft::block_cipher_trait::BlockCipher;