+++ src/prelude.rs
@@ -2,11 +2,11 @@
 pub use crate::traits::wg::prelude::*;
 
 // Uhh get rid of this as soon as Vadim did v3
-pub use crate::traits::wg::digital::v2::OutputPin as _embedded_hal_digital_v2_OutputPin;
-pub use crate::traits::wg::digital::v2::StatefulOutputPin as _embedded_hal_digital_v2_StatefulOutputPin;
+pub use crate::traits::wg::digital::v2::OutputPin as _;
+pub use crate::traits::wg::digital::v2::StatefulOutputPin as _;
 
 // pub use crate::time::Sleep as _lpc55s6x_hal_time_Sleep;
-pub use crate::time::U32Ext as _lpc55s6x_hal_time_U32Ext;
+pub use crate::time::U32Ext as _;
 
 pub use crate::traits::flash::Read as _;
 pub use crate::traits::flash::WriteErase as _;